Got the chatbot to connect and read messages
Not sure why imports were dying but when I made the config a class and
imported the class it worked. Found a better way to parge twitch
messages with a compiled regex which just returns a string which made it
smoother.

I ran it and tested it by typing things in the channel chat and it was
reading it !!!